chore(fiters): Move components into folders
