+++ projects/hslayers/common/filters/comparison-filter/comparison-filter.component.scss
