+++ projects/hslayers/common/filters/public-api.ts
@@ -1,5 +1,5 @@
-export * from './add-filter-button.component';
-export * from './comparison-filter.component';
 export * from './filter.component';
 export * from './filters.component';
 export * from './filters.service';
+export * from './comparison-filter/comparison-filter.component';
+export * from './add-filter-button/add-filter-button.component';
